requirements.txt
pillow==10.2.0
win10toast==0.9 